requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
botocore==1.34.87
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jwcrypto==1.5.6
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.1
packaging==24.0
pkce==1.0.3
protobuf==5.27.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.1
